Added stream=True to HorridAPI chat completions to enable real-time responses
This commit updates the HorridAPI chat completions to enable real-time responses by setting stream=True. This allows the API to send responses as they are generated, rather than waiting for the entire response to be completed.

When stream=True is set, the API will send a stream of responses, which can be useful for applications that require real-time interactions. For example:


from HorridAPI import Mango

mango = Mango()

response = mango.chat.completions.create(
    model="gpt-3.5",
    stream=True,
    messages=[{"role": "user", "content": "Hello"}]
)

print(response.text)


This will output:


Hi
There!
How
Can
I
Help
You


If stream=True is not set, the API will wait for the entire response to be completed before sending it, which can result in slower response times